feat: Add assembly code for calculating factorial
This commit adds the assembly code for calculating the factorial of a number. The code is implemented in the `fat.asm` file and includes the necessary instructions to take user input, perform the factorial calculation, and output the result.

The `fat.pre` file is also added, which contains the pre-processed version of the assembly code.

The purpose of this commit is to provide the initial implementation of the factorial calculation functionality in the Montador project.